Create Boolean_minimizer.py
A Boolean function minimizer is used in optimizing of logic circuits. The problem posed by this
project is the development of a software that would enable the reduction of a Boolean function of
up to 8 variable functions. The result of this reduction, that is, the minimized function is to be
displayed together with its implementation using logic gates.